+++ src/features/Counter/lib/utils/demarhsallMinMaxValues.ts
@@ -1,6 +1,6 @@
 import { MinMaxValues } from '../types/counter.types';
 import { not } from '@/common/utils/not';
-import { fieldValuesValidator } from './fieldValueValidator';
+import { fieldValuesValidator } from './fieldValuesValidator';
 
 export const demarshallMinMaxValues = (
     storedValues: string | null,
