+++ src/features/Counter/lib/validators/fieldValuesValidator/validator/validators/max_gt_Min.ts
@@ -1,12 +1,12 @@
 import { IncorrectFieldName } from '@/features/Counter/lib/types/counter.types';
 import { Validator } from '../validator';
 
-export class ValidateBoth extends Validator {
+export class Max_gt_Min extends Validator {
     constructor(incFieldName: IncorrectFieldName, errorText: string) {
         super(incFieldName, errorText);
     }
 
     validateValues(minValue: number, maxValue: number): boolean {
-        return minValue !== maxValue;
+        return minValue < maxValue;
     }
 }
