+++ src/features/Counter/lib/validators/fieldValuesValidator/validator/validators/valuesAreNotEqual.ts
@@ -1,12 +1,12 @@
 import { IncorrectFieldName } from '@/features/Counter/lib/types/counter.types';
 import { Validator } from '../validator';
 
-export class ValidateMax extends Validator {
+export class ValuesAreNotEqual extends Validator {
     constructor(incFieldName: IncorrectFieldName, errorText: string) {
         super(incFieldName, errorText);
     }
 
     validateValues(minValue: number, maxValue: number): boolean {
-        return minValue < maxValue;
+        return minValue !== maxValue;
     }
 }
